+++ CHANGELOG.md
@@ -81,11 +81,13 @@
     - Added ```Controller.sdkVersion``` property
     - Added ```Controller.firmwareVersion``` property
     - Added ```Controller.HomeIDAsHex``` property
+    - Added ```ZWaveNode.InterviewStageCompleted``` event
     - Implemented ```INotifyPropertyChanged``` support for the below.  
     this improves support for MVVM and allows UI bindings to update automatically.
         - ```ZWaveNode.statistics```
         - ```ZWaveNode.status```
         - ```ZWaveNode.ready```
+        - ```ZWaveNode.interviewStage```
         - ```Controller.statistics```
         - ```Controller.isRebuildingRoutes```